Back_vecinApp/requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
mysqlclient==2.2.8
pillow==12.1.1
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
